Pulled block functions into new block reducer
